+++ bank-feeds/src/codat_bankfeeds/source_accounts.py
@@ -6,7 +6,7 @@ from codat_bankfeeds._hooks import HookContext
 from codat_bankfeeds.models import errors, operations, shared
 from codat_bankfeeds.types import BaseModel, OptionalNullable, UNSET
 import io
-from typing import Any, IO, Mapping, Optional, Union, cast
+from typing import Any, IO, List, Mapping, Optional, Union, cast
 
 
 class SourceAccounts(BaseSDK):
@@ -366,28 +366,26 @@ class SourceAccounts(BaseSDK):
         data: Any = None
         if utils.match_response(http_res, "201", "application/json"):
             return utils.unmarshal_json(
-                http_res.text, Optional[operations.CreateBatchSourceAccountResponseBody]
+                http_res.text, Optional[List[operations.ResponseBody]]
             )
         if utils.match_response(http_res, "207", "application/json"):
             return utils.unmarshal_json(
                 http_res.text,
-                Optional[operations.CreateBatchSourceAccountSourceAccountsResponseBody],
+                Optional[List[operations.CreateBatchSourceAccountResponseBody]],
             )
         if utils.match_response(
-            http_res,
-            ["400", "401", "402", "403", "404", "409", "429"],
-            "application/json",
+            http_res, ["400", "401", "402", "403", "404", "429"], "application/json"
         ):
             data = utils.unmarshal_json(http_res.text, errors.ErrorMessageData)
             raise errors.ErrorMessage(data=data)
-        if utils.match_response(http_res, ["500", "503"], "application/json"):
-            data = utils.unmarshal_json(http_res.text, errors.ErrorMessageData)
-            raise errors.ErrorMessage(data=data)
-        if utils.match_response(http_res, "4XX", "*"):
+        if utils.match_response(http_res, ["409", "4XX"], "*"):
             http_res_text = utils.stream_to_text(http_res)
             raise errors.SDKError(
                 "API error occurred", http_res.status_code, http_res_text, http_res
             )
+        if utils.match_response(http_res, ["500", "503"], "application/json"):
+            data = utils.unmarshal_json(http_res.text, errors.ErrorMessageData)
+            raise errors.ErrorMessage(data=data)
         if utils.match_response(http_res, "5XX", "*"):
             http_res_text = utils.stream_to_text(http_res)
             raise errors.SDKError(
@@ -503,28 +501,26 @@ class SourceAccounts(BaseSDK):
         data: Any = None
         if utils.match_response(http_res, "201", "application/json"):
             return utils.unmarshal_json(
-                http_res.text, Optional[operations.CreateBatchSourceAccountResponseBody]
+                http_res.text, Optional[List[operations.ResponseBody]]
             )
         if utils.match_response(http_res, "207", "application/json"):
             return utils.unmarshal_json(
                 http_res.text,
-                Optional[operations.CreateBatchSourceAccountSourceAccountsResponseBody],
+                Optional[List[operations.CreateBatchSourceAccountResponseBody]],
             )
         if utils.match_response(
-            http_res,
-            ["400", "401", "402", "403", "404", "409", "429"],
-            "application/json",
+            http_res, ["400", "401", "402", "403", "404", "429"], "application/json"
         ):
             data = utils.unmarshal_json(http_res.text, errors.ErrorMessageData)
             raise errors.ErrorMessage(data=data)
-        if utils.match_response(http_res, ["500", "503"], "application/json"):
-            data = utils.unmarshal_json(http_res.text, errors.ErrorMessageData)
-            raise errors.ErrorMessage(data=data)
-        if utils.match_response(http_res, "4XX", "*"):
+        if utils.match_response(http_res, ["409", "4XX"], "*"):
             http_res_text = await utils.stream_to_text_async(http_res)
             raise errors.SDKError(
                 "API error occurred", http_res.status_code, http_res_text, http_res
             )
+        if utils.match_response(http_res, ["500", "503"], "application/json"):
+            data = utils.unmarshal_json(http_res.text, errors.ErrorMessageData)
+            raise errors.ErrorMessage(data=data)
         if utils.match_response(http_res, "5XX", "*"):
             http_res_text = await utils.stream_to_text_async(http_res)
             raise errors.SDKError(